requirements.txt
alembic==1.6.5
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
blis==0.7.4
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cymem==2.0.5
dnspython==2.1.0
elasticsearch==7.13.1
email-validator==1.1.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Admin==1.5.8
Flask-Babel==2.0.0
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==3.0.1
Flask-Moment==1.0.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
Mako==1.1.4
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
murmurhash==1.0.5
nltk==3.6.3
numpy==1.21.2
openpyxl==3.0.9
packaging==21.0
pandas==1.3.4
passlib==1.7.4
pathy==0.6.0
Pillow==8.3.1
preshed==3.0.5
psycopg2-binary==2.9.1
pugsql==0.2.3
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
speaklater==1.3
SQLAlchemy==1.4.17
srsly==2.4.1
tablib==3.0.0
thinc==8.0.10
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
WTForms==2.3.3
xlrd==2.0.1
xlwt==1.3.0
